requirements.txt
GitPython==3.1.29
matplotlib==3.5.2
matplotlib-inline==0.1.3
numpy==1.22.3
scikit-learn>=0.20
seaborn==0.11.2
